separate CLI tests
